Some test cases
